refactor: reduce redundancy for AccountDTOs
