+++ src/main/java/io/github/janmalch/kino/api/boundary/AccountResource.java
@@ -3,7 +3,7 @@ package io.github.janmalch.kino.api.boundary;
 import io.github.janmalch.kino.api.ResponseResultBuilder;
 import io.github.janmalch.kino.api.SuccessMessage;
 import io.github.janmalch.kino.api.model.account.AccountInfoDto;
-import io.github.janmalch.kino.api.model.auth.SignUpDto;
+import io.github.janmalch.kino.api.model.account.SignUpDto;
 import io.github.janmalch.kino.control.account.*;
 import io.github.janmalch.kino.security.Secured;
 import io.swagger.annotations.Api;
