+++ src/main/java/io/github/janmalch/kino/api/boundary/MyAccountResource.java
@@ -2,7 +2,7 @@ package io.github.janmalch.kino.api.boundary;
 
 import io.github.janmalch.kino.api.ResponseResultBuilder;
 import io.github.janmalch.kino.api.model.account.AccountInfoDto;
-import io.github.janmalch.kino.api.model.auth.SignUpDto;
+import io.github.janmalch.kino.api.model.account.SignUpDto;
 import io.github.janmalch.kino.api.model.auth.TokenDto;
 import io.github.janmalch.kino.control.myaccount.DeleteMyAccountControl;
 import io.github.janmalch.kino.control.myaccount.EditMyAccountControl;
