+++ src/main/java/io/github/janmalch/kino/api/model/account/AbstractAccountDto.java
@@ -1,0 +1,32 @@
+package io.github.janmalch.kino.api.model.account;
+
+public abstract class AbstractAccountDto {
+
+  protected String firstName;
+  protected String lastName;
+  protected String email;
+
+  public String getFirstName() {
+    return firstName;
+  }
+
+  public void setFirstName(String firstName) {
+    this.firstName = firstName;
+  }
+
+  public String getLastName() {
+    return lastName;
+  }
+
+  public void setLastName(String lastName) {
+    this.lastName = lastName;
+  }
+
+  public String getEmail() {
+    return email;
+  }
+
+  public void setEmail(String email) {
+    this.email = email;
+  }
+}
