+++ src/main/java/io/github/janmalch/kino/api/model/account/AccountInfoDto.java
@@ -1,6 +1,5 @@
 package io.github.janmalch.kino.api.model.account;
 
-import io.github.janmalch.kino.api.model.auth.SignUpDto;
 import io.github.janmalch.kino.entity.Role;
 
 public class AccountInfoDto extends SignUpDto {
