+++ src/main/java/io/github/janmalch/kino/api/model/account/SafeAccountInfoDto.java
@@ -2,12 +2,9 @@ package io.github.janmalch.kino.api.model.account;
 
 import io.github.janmalch.kino.entity.Role;
 
-public class SafeAccountInfoDto {
+public class SafeAccountInfoDto extends AbstractAccountDto {
 
   private long id;
-  private String firstName;
-  private String lastName;
-  private String email;
   private Role role;
 
   public long getId() {
@@ -16,30 +13,6 @@ public class SafeAccountInfoDto {
 
   public void setId(long id) {
     this.id = id;
-  }
-
-  public String getFirstName() {
-    return firstName;
-  }
-
-  public void setFirstName(String firstName) {
-    this.firstName = firstName;
-  }
-
-  public String getLastName() {
-    return lastName;
-  }
-
-  public void setLastName(String lastName) {
-    this.lastName = lastName;
-  }
-
-  public String getEmail() {
-    return email;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
   }
 
   public Role getRole() {
