+++ src/test/java/io/github/janmalch/kino/api/boundary/AccountResourceTest.java
@@ -3,7 +3,7 @@ package io.github.janmalch.kino.api.boundary;
 import static org.junit.jupiter.api.Assertions.*;
 
 import io.github.janmalch.kino.api.model.account.AccountInfoDto;
-import io.github.janmalch.kino.api.model.auth.SignUpDto;
+import io.github.janmalch.kino.api.model.account.SignUpDto;
 import io.github.janmalch.kino.entity.Account;
 import io.github.janmalch.kino.entity.EntityWiper;
 import io.github.janmalch.kino.entity.Role;
