+++ src/test/java/io/github/janmalch/kino/api/boundary/AuthResourceTest.java
@@ -2,8 +2,8 @@ package io.github.janmalch.kino.api.boundary;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import io.github.janmalch.kino.api.model.account.SignUpDto;
 import io.github.janmalch.kino.api.model.auth.LoginDto;
-import io.github.janmalch.kino.api.model.auth.SignUpDto;
 import io.github.janmalch.kino.api.model.auth.TokenDto;
 import io.github.janmalch.kino.entity.Account;
 import io.github.janmalch.kino.repository.Repository;
