+++ src/test/java/io/github/janmalch/kino/control/validation/SignUpDtoValidatorTest.java
@@ -2,7 +2,7 @@ package io.github.janmalch.kino.control.validation;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import io.github.janmalch.kino.api.model.auth.SignUpDto;
+import io.github.janmalch.kino.api.model.account.SignUpDto;
 import io.github.janmalch.kino.problem.Problem;
 import java.time.LocalDate;
 import java.util.Optional;
